+++ NEWS.rst
@@ -4,6 +4,8 @@ NEWS for system-image updater
 
 1.1 (2013-XX-XX)
 ================
+ * Use nose as the test runner.  This allows us to pre-initialize the logging
+   to prevent unwanted output. (LP: #1207117)
 
 1.0 (2013-08-01)
 ================